requirements.txt
# Core dependencies
pyspark==3.5.0
delta-spark==3.0.0

# Azure SDK
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Data quality & testing
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
